Conflicts:
	configs/brick_master_config.py

+++ python/generate_python_zip.py
@@ -55,13 +55,6 @@ def make_files(com_new, directory):
     copy_examples_for_zip()
 
 def generate(path):
-    common.path_binding = path
-    path_list = path.split('/')
-    path_list[-1] = 'configs'
-    path_config = '/'.join(path_list)
-    sys.path.append(path_config)
-    configs = os.listdir(path_config)
-
     # Make temporary generator directory
     if os.path.exists('/tmp/generator'):
         shutil.rmtree('/tmp/generator/')
@@ -69,7 +62,7 @@ def generate(path):
     os.chdir('/tmp/generator')
 
     # Copy examples
-    common.import_and_make(configs, path, make_files)
+    common.generate(path, 'en', make_files, None, False)
     shutil.copy(common.path_binding.replace('/generators/python', '/doc/en/source/Software/example.py'),
                 '/tmp/generator/egg/examples/example_enumerate.py')
 
@@ -102,7 +95,8 @@ setup(name='tinkerforge',
     args = ['/usr/bin/python',
             'setup.py',
             'bdist_egg']
-    subprocess.call(args)
+    if subprocess.call(args) != 0:
+        raise Exception("Command '{0}' failed".format(' '.join(args)))
 
     # Remove build stuff
     shutil.rmtree('/tmp/generator/egg/source/build')
